Fix compatibility with Android 12
