+++ app/src/main/java/org/lsposed/manager/ui/activity/MainActivity.java
@@ -27,7 +27,6 @@ import android.text.TextUtils;
 import android.view.KeyEvent;
 import android.view.MotionEvent;
 import android.view.View;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.core.os.BuildCompat;
